apps/agent/requirements.txt
livekit-agents
livekit-plugins-openai
httpx
